+++ internal/runners/update/update.go
@@ -3,14 +3,18 @@ package update
 import (
 	"context"
 	"os"
+	"path/filepath"
 
+	"github.com/ActiveState/cli/internal/appinfo"
 	"github.com/ActiveState/cli/internal/config"
 	"github.com/ActiveState/cli/internal/constants"
 	"github.com/ActiveState/cli/internal/errs"
+	"github.com/ActiveState/cli/internal/installation"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/output"
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/prompt"
+	"github.com/ActiveState/cli/internal/svcmanager"
 	"github.com/ActiveState/cli/pkg/platform/model"
 	"github.com/ActiveState/cli/pkg/project"
 )
@@ -23,6 +27,7 @@ type Params struct {
 type Update struct {
 	project *project.Project
 	cfg     *config.Instance
+	svcmgr  *svcmanager.Manager
 	out     output.Outputer
 	prompt  prompt.Prompter
 }
@@ -30,6 +35,7 @@ type Update struct {
 type primeable interface {
 	primer.Projecter
 	primer.Configurer
+	primer.Svcer
 	primer.Outputer
 	primer.Prompter
 }
@@ -38,17 +44,22 @@ func New(prime primeable) *Update {
 	return &Update{
 		prime.Project(),
 		prime.Config(),
+		prime.SvcManager(),
 		prime.Output(),
 		prime.Prompt(),
 	}
 }
 
 func (u *Update) Run(params *Params) error {
-	u.out.Notice(locale.Tl("updating_version", "Updating State Tool to latest version available."))
+	if params.Version == "" {
+		u.out.Notice(locale.Tl("updating_latest", "Updating State Tool to latest version available."))
+	} else {
+		u.out.Notice(locale.Tl("updating_version", "Updating State Tool to version {{.V0}}", params.Version))
+	}
 
 	channel := fetchChannel(params.Channel, true)
 
-	m, err := model.NewSvcModel(context.Background(), u.cfg)
+	m, err := model.NewSvcModel(context.Background(), u.cfg, u.svcmgr)
 	if err != nil {
 		return errs.Wrap(err, "failed to create svc model")
 	}
@@ -62,6 +73,16 @@ func (u *Update) Run(params *Params) error {
 	if up.Channel == "" && up.Version == "" {
 		u.out.Print(locale.Tl("update_uptodate", "You are already using the latest State Tool version available."))
 		return nil
+	}
+
+	// Stop currently running applications (state-tray and state-svc) if we are switching channels.
+	// When we switch channels the config directory changes and the deferred update cannot stop the
+	// running applications.
+	if up.Channel != constants.BranchName {
+		err = installation.StopRunning(filepath.Dir(appinfo.StateApp().Exec()))
+		if err != nil {
+			return errs.Wrap(err, "Could not stop running services")
+		}
 	}
 
 	u.out.Print(locale.Tl("version_updating_deferred", "Version update to {{.V0}}@{{.V1}} has started and should complete in seconds.\nRefer to log file [ACTIONABLE]{{.V2}}[/RESET] for progress.", up.Channel, up.Version, up.Logfile))