+++ internal/subshell/cmd/cmd.go
@@ -86,7 +86,30 @@ func (v *SubShell) WriteUserEnv(cfg sscommon.Configurable, env map[string]string
 		}
 	}
 
-	osutils.PropagateEnv()
+	if err := osutils.PropagateEnv(); err != nil {
+		return errs.Wrap(err, "Sending OS signal to update environment failed.")
+	}
+	return nil
+}
+
+func (v *SubShell) CleanUserEnv(cfg sscommon.Configurable, envType sscommon.RcIdentification, userScope bool) error {
+	cmdEnv := NewCmdEnv(userScope)
+
+	// Clean up old entries
+	oldEnv := cfg.GetStringMap(envType.Key)
+	for k, v := range oldEnv {
+		if err := cmdEnv.unset(k, v.(string)); err != nil {
+			return err
+		}
+	}
+
+	if err := osutils.PropagateEnv(); err != nil {
+		return errs.Wrap(err, "Sending OS signal to update environment failed.")
+	}
+	return nil
+}
+
+func (v *SubShell) RemoveLegacyInstallPath(_ sscommon.Configurable) error {
 	return nil
 }